pyproject.toml
[project]
name = "parmair-ha"
version = "0.9.0"
description = "Home Assistant integration for Parmair MAC ventilation units"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "ValtteriAho"}
]
keywords = ["home-assistant", "parmair", "ventilation", "modbus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pymodbus>=3.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "ruff==0.15.1",  # match .pre-commit-config.yaml rev so local format matches hook
    "mypy>=1.10.0",
    "homeassistant-stubs>=2025.5.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/ValtteriAho/Hassio_ParmAir"
Repository = "https://github.com/ValtteriAho/Hassio_ParmAir"
Issues = "https://github.com/ValtteriAho/Hassio_ParmAir/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["custom_components*", "tools*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["custom_components/parmair", "tools"]
omit = ["*/__pycache__/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
no_implicit_optional = true
strict_equality = true
show_error_codes = true

# Disable checks that are too noisy for this codebase
warn_return_any = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "pymodbus.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "ARG002", # unused method argument - required by Home Assistant API signatures
]

[tool.ruff.lint.isort]
known-first-party = ["custom_components.parmair"]

